+++ packages/react-internal/src/components/TextField/MaskedTextField/MaskedTextField.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { TextField } from '../TextField';
-import { ITextField, IMaskedTextFieldProps } from '../TextField.types';
-import { initializeComponentRef, KeyCodes } from '../../../Utilities';
-
+import { IMaskedTextFieldProps, IMaskedTextField } from '../TextField.types';
+import { KeyCodes, IRefObject } from '../../../Utilities';
 import {
   clearNext,
   clearPrev,
@@ -15,386 +14,364 @@ import {
   insertString,
   parseMask,
 } from './inputMask';
+import { useConst } from '@uifabric/react-hooks';
 
-/**
- * State for the MaskedTextField component.
- */
-export interface IMaskedTextFieldState {
+interface IMaskedTextFieldInternalState {
+  maskCharData: IMaskValue[];
+  isFocused: boolean;
+  moveCursorOnMouseUp: boolean;
+  changeSelectionData: {
+    changeType: 'default' | 'backspace' | 'delete' | 'textPasted';
+    selectionStart: number;
+    selectionEnd: number;
+  } | null;
+}
+
+const COMPONENT_NAME = 'MaskedTextField';
+
+const useComponentRef = (
+  componentRef: IRefObject<IMaskedTextField> | undefined,
+  internalState: IMaskedTextFieldInternalState,
+  textField: React.RefObject<IMaskedTextField>,
+) => {
+  React.useImperativeHandle(
+    componentRef,
+    (): IMaskedTextField => ({
+      get value() {
+        let value = '';
+
+        for (let i = 0; i < internalState.maskCharData.length; i++) {
+          if (!internalState.maskCharData[i].value) {
+            return undefined;
+          }
+          value += internalState.maskCharData[i].value;
+        }
+        return value;
+      },
+
+      get selectionStart(): number | null {
+        return textField.current && textField.current.selectionStart !== null ? textField.current.selectionStart : -1;
+      },
+
+      get selectionEnd(): number | null {
+        return textField.current && textField.current.selectionEnd ? textField.current.selectionEnd : -1;
+      },
+
+      focus(): void {
+        textField.current && textField.current.focus();
+      },
+
+      blur(): void {
+        textField.current && textField.current.blur();
+      },
+
+      select(): void {
+        textField.current && textField.current.select();
+      },
+
+      setSelectionStart(value: number): void {
+        textField.current && textField.current.setSelectionStart(value);
+      },
+
+      setSelectionEnd(value: number): void {
+        textField.current && textField.current.setSelectionEnd(value);
+      },
+
+      setSelectionRange(start: number, end: number): void {
+        textField.current && textField.current.setSelectionRange(start, end);
+      },
+    }),
+    [internalState, textField],
+  );
+};
+
+export const DEFAULT_MASK_CHAR = '_';
+
+export const MaskedTextField = React.forwardRef<HTMLDivElement, IMaskedTextFieldProps>((props, ref) => {
+  const textField = React.useRef<IMaskedTextField>(null);
+
+  const {
+    componentRef,
+    onFocus,
+    onBlur,
+    onMouseDown,
+    onMouseUp,
+    onChange,
+    onPaste,
+    onKeyDown,
+    mask,
+    maskChar = DEFAULT_MASK_CHAR,
+    maskFormat = DEFAULT_MASK_FORMAT_CHARS,
+    value,
+  } = props;
+
+  const internalState = useConst<IMaskedTextFieldInternalState>(() => ({
+    maskCharData: parseMask(mask, maskFormat),
+    isFocused: false,
+    moveCursorOnMouseUp: false,
+    changeSelectionData: null,
+  }));
+
+  /** The index into the rendered value of the first unfilled format character */
+  const [maskCursorPosition, setMaskCursorPosition] = React.useState<number | undefined>();
+
   /**
    * The mask string formatted with the input value.
    * This is what is displayed inside the TextField
    * @example
    *  `Phone Number: 12_ - 4___`
    */
-  displayValue: string;
-  /** The index into the rendered value of the first unfilled format character */
-  maskCursorPosition?: number;
-}
+  const [displayValue, setDisplayValue] = React.useState<string>(() =>
+    getMaskDisplay(mask, internalState.maskCharData, maskChar),
+  );
 
-export const DEFAULT_MASK_CHAR = '_';
+  const setValue = React.useCallback(
+    (newValue: string): void => {
+      let valueIndex = 0;
+      let charDataIndex = 0;
 
-type InputChangeType = 'default' | 'backspace' | 'delete' | 'textPasted';
-
-export class MaskedTextField extends React.Component<IMaskedTextFieldProps, IMaskedTextFieldState>
-  implements ITextField {
-  public static defaultProps: IMaskedTextFieldProps = {
-    maskChar: DEFAULT_MASK_CHAR,
-    maskFormat: DEFAULT_MASK_FORMAT_CHARS,
-  };
-
-  private _textField = React.createRef<ITextField>();
-
-  /**
-   *  An array of data containing information regarding the format characters,
-   *  their indices inside the display text, and their corresponding values.
-   * @example
-   * ```
-   *  [
-   *    { value: '1', displayIndex: 16, format: /[0-9]/ },
-   *    { value: '2', displayIndex: 17, format: /[0-9]/ },
-   *    { displayIndex: 18, format: /[0-9]/ },
-   *    { value: '4', displayIndex: 22, format: /[0-9]/ },
-   *    ...
-   *  ]
-   * ```
-   */
-  private _maskCharData: IMaskValue[];
-  /** True if the TextField is focused */
-  private _isFocused: boolean;
-  /** True if the TextField was not focused and it was clicked into */
-  private _moveCursorOnMouseUp: boolean;
-
-  /** The stored selection data prior to input change events. */
-  private _changeSelectionData: {
-    changeType: InputChangeType;
-    selectionStart: number;
-    selectionEnd: number;
-  } | null;
-
-  constructor(props: IMaskedTextFieldProps) {
-    super(props);
-
-    initializeComponentRef(this);
-
-    // Translate mask into charData
-    this._maskCharData = parseMask(props.mask, props.maskFormat);
-    // If an initial value is provided, use it to populate the format chars
-    props.value !== undefined && this.setValue(props.value);
-
-    this._isFocused = false;
-    this._moveCursorOnMouseUp = false;
-
-    this.state = {
-      displayValue: getMaskDisplay(props.mask, this._maskCharData, props.maskChar),
-    };
-  }
-
-  public UNSAFE_componentWillReceiveProps(newProps: IMaskedTextFieldProps) {
-    if (newProps.mask !== this.props.mask || newProps.value !== this.props.value) {
-      this._maskCharData = parseMask(newProps.mask, newProps.maskFormat);
-      newProps.value !== undefined && this.setValue(newProps.value);
-
-      this.setState({
-        displayValue: getMaskDisplay(newProps.mask, this._maskCharData, newProps.maskChar),
-      });
-    }
-  }
-
-  public componentDidUpdate() {
-    // Move the cursor to the start of the mask format on update
-    if (this._isFocused && this.state.maskCursorPosition !== undefined && this._textField.current) {
-      this._textField.current.setSelectionRange(this.state.maskCursorPosition, this.state.maskCursorPosition);
-    }
-  }
-
-  public render() {
-    return (
-      <TextField
-        {...this.props}
-        onFocus={this._onFocus}
-        onBlur={this._onBlur}
-        onMouseDown={this._onMouseDown}
-        onMouseUp={this._onMouseUp}
-        onChange={this._onInputChange}
-        onKeyDown={this._onKeyDown}
-        onPaste={this._onPaste}
-        value={this.state.displayValue || ''}
-        componentRef={this._textField}
-      />
-    );
-  }
-
-  /**
-   * @returns The value of all filled format characters or undefined if not all format characters are filled
-   */
-  public get value(): string | undefined {
-    let value = '';
-
-    for (let i = 0; i < this._maskCharData.length; i++) {
-      if (!this._maskCharData[i].value) {
-        return undefined;
+      while (valueIndex < newValue.length && charDataIndex < internalState.maskCharData.length) {
+        // Test if the next character in the new value fits the next format character
+        const testVal = newValue[valueIndex];
+        if (internalState.maskCharData[charDataIndex].format.test(testVal)) {
+          internalState.maskCharData[charDataIndex].value = testVal;
+          charDataIndex++;
+        }
+        valueIndex++;
       }
-      value += this._maskCharData[i].value;
-    }
-    return value;
-  }
+    },
+    [internalState],
+  );
 
-  public setValue(newValue: string): void {
-    let valueIndex = 0;
-    let charDataIndex = 0;
+  const handleFocus = React.useCallback(
+    (ev: React.FocusEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+      onFocus?.(ev);
+      internalState.isFocused = true;
 
-    while (valueIndex < newValue.length && charDataIndex < this._maskCharData.length) {
-      // Test if the next character in the new value fits the next format character
-      const testVal = newValue[valueIndex];
-      if (this._maskCharData[charDataIndex].format.test(testVal)) {
-        this._maskCharData[charDataIndex].value = testVal;
-        charDataIndex++;
-      }
-      valueIndex++;
-    }
-  }
-
-  public focus(): void {
-    const { current } = this._textField;
-
-    current && current.focus();
-  }
-
-  public blur(): void {
-    const { current } = this._textField;
-
-    current && current.blur();
-  }
-
-  public select(): void {
-    const { current } = this._textField;
-
-    current && current.select();
-  }
-
-  public setSelectionStart(value: number): void {
-    const { current } = this._textField;
-
-    current && current.setSelectionStart(value);
-  }
-
-  public setSelectionEnd(value: number): void {
-    const { current } = this._textField;
-
-    current && current.setSelectionEnd(value);
-  }
-
-  public setSelectionRange(start: number, end: number): void {
-    const { current } = this._textField;
-    current && current.setSelectionRange(start, end);
-  }
-
-  public get selectionStart(): number | null {
-    const { current } = this._textField;
-
-    return current && current.selectionStart !== null ? current.selectionStart : -1;
-  }
-
-  public get selectionEnd(): number | null {
-    const { current } = this._textField;
-
-    return current && current.selectionEnd ? current.selectionEnd : -1;
-  }
-
-  private _onFocus = (event: React.FocusEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    if (this.props.onFocus) {
-      this.props.onFocus(event);
-    }
-
-    this._isFocused = true;
-
-    // Move the cursor position to the leftmost unfilled position
-    for (let i = 0; i < this._maskCharData.length; i++) {
-      if (!this._maskCharData[i].value) {
-        this.setState({
-          maskCursorPosition: this._maskCharData[i].displayIndex,
-        });
-        break;
-      }
-    }
-  };
-
-  private _onBlur = (event: React.FocusEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    if (this.props.onBlur) {
-      this.props.onBlur(event);
-    }
-
-    this._isFocused = false;
-    this._moveCursorOnMouseUp = true;
-  };
-
-  private _onMouseDown = (event: React.MouseEvent<HTMLInputElement>) => {
-    if (this.props.onMouseDown) {
-      this.props.onMouseDown(event);
-    }
-
-    if (!this._isFocused) {
-      this._moveCursorOnMouseUp = true;
-    }
-  };
-
-  private _onMouseUp = (event: React.MouseEvent<HTMLInputElement>) => {
-    if (this.props.onMouseUp) {
-      this.props.onMouseUp(event);
-    }
-
-    // Move the cursor on mouseUp after focusing the textField
-    if (this._moveCursorOnMouseUp) {
-      this._moveCursorOnMouseUp = false;
-      // Move the cursor position to the rightmost unfilled position
-      for (let i = 0; i < this._maskCharData.length; i++) {
-        if (!this._maskCharData[i].value) {
-          this.setState({
-            maskCursorPosition: this._maskCharData[i].displayIndex,
-          });
+      // Move the cursor position to the leftmost unfilled position
+      for (let i = 0; i < internalState.maskCharData.length; i++) {
+        if (!internalState.maskCharData[i].value) {
+          setMaskCursorPosition(internalState.maskCharData[i].displayIndex);
           break;
         }
       }
-    }
-  };
+    },
+    [internalState, onFocus],
+  );
 
-  private _onInputChange = (ev: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>, value: string) => {
-    const textField = this._textField.current;
-    if (this._changeSelectionData === null && textField) {
-      this._changeSelectionData = {
-        changeType: 'default',
-        selectionStart: textField.selectionStart !== null ? textField.selectionStart : -1,
-        selectionEnd: textField.selectionEnd !== null ? textField.selectionEnd : -1,
-      };
-    }
-    if (!this._changeSelectionData) {
-      return;
-    }
+  const handleBlur = React.useCallback(
+    (ev: React.FocusEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+      onBlur?.(ev);
+      internalState.isFocused = false;
+      internalState.moveCursorOnMouseUp = true;
+    },
+    [internalState, onBlur],
+  );
 
-    const { displayValue } = this.state;
-
-    // The initial value of cursorPos does not matter
-    let cursorPos = 0;
-    const { changeType, selectionStart, selectionEnd } = this._changeSelectionData;
-
-    if (changeType === 'textPasted') {
-      const charsSelected = selectionEnd - selectionStart;
-      const charCount = value.length + charsSelected - displayValue.length;
-      const startPos = selectionStart;
-      const pastedString = value.substr(startPos, charCount);
-
-      // Clear any selected characters
-      if (charsSelected) {
-        this._maskCharData = clearRange(this._maskCharData, selectionStart, charsSelected);
+  const handleMouseDown = React.useCallback(
+    (ev: React.MouseEvent<HTMLInputElement>) => {
+      onMouseDown?.(ev);
+      if (!internalState.isFocused) {
+        internalState.moveCursorOnMouseUp = true;
       }
-      cursorPos = insertString(this._maskCharData, startPos, pastedString);
-    } else if (changeType === 'delete' || changeType === 'backspace') {
-      // isDel is true If the characters are removed LTR, otherwise RTL
-      const isDel = changeType === 'delete';
-      const charCount = selectionEnd - selectionStart;
+    },
+    [internalState, onMouseDown],
+  );
 
-      if (charCount) {
-        // charCount is > 0 if range was deleted
-        this._maskCharData = clearRange(this._maskCharData, selectionStart, charCount);
-        cursorPos = getRightFormatIndex(this._maskCharData, selectionStart);
-      } else {
-        // If charCount === 0, there was no selection and a single character was deleted
-        if (isDel) {
-          this._maskCharData = clearNext(this._maskCharData, selectionStart);
-          cursorPos = getRightFormatIndex(this._maskCharData, selectionStart);
-        } else {
-          this._maskCharData = clearPrev(this._maskCharData, selectionStart);
-          cursorPos = getLeftFormatIndex(this._maskCharData, selectionStart);
+  const handleMouseUp = React.useCallback(
+    (ev: React.MouseEvent<HTMLInputElement>) => {
+      onMouseUp?.(ev);
+      // Move the cursor on mouseUp after focusing the textField
+      if (internalState.moveCursorOnMouseUp) {
+        internalState.moveCursorOnMouseUp = false;
+        // Move the cursor position to the rightmost unfilled position
+        for (let i = 0; i < internalState.maskCharData.length; i++) {
+          if (!internalState.maskCharData[i].value) {
+            setMaskCursorPosition(internalState.maskCharData[i].displayIndex);
+            break;
+          }
         }
       }
-    } else if (value.length > displayValue.length) {
-      // This case is if the user added characters
-      const charCount = value.length - displayValue.length;
-      const startPos = selectionEnd - charCount;
-      const enteredString = value.substr(startPos, charCount);
+    },
+    [internalState, onMouseUp],
+  );
 
-      cursorPos = insertString(this._maskCharData, startPos, enteredString);
-    } else if (value.length <= displayValue.length) {
-      /**
-       * This case is reached only if the user has selected a block of 1 or more
-       * characters and input a character replacing the characters they've selected.
-       */
-      const charCount = 1;
-      const selectCount = displayValue.length + charCount - value.length;
-      const startPos = selectionEnd - charCount;
-      const enteredString = value.substr(startPos, charCount);
-
-      // Clear the selected range
-      this._maskCharData = clearRange(this._maskCharData, startPos, selectCount);
-      // Insert the printed character
-      cursorPos = insertString(this._maskCharData, startPos, enteredString);
-    }
-
-    this._changeSelectionData = null;
-
-    const newValue = getMaskDisplay(this.props.mask, this._maskCharData, this.props.maskChar);
-
-    this.setState({
-      displayValue: newValue,
-      maskCursorPosition: cursorPos,
-    });
-
-    // Perform onChange after input has been processed. Return value is expected to be the displayed text
-    if (this.props.onChange) {
-      this.props.onChange(ev, newValue);
-    }
-  };
-
-  private _onKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    const { current } = this._textField;
-
-    if (this.props.onKeyDown) {
-      this.props.onKeyDown(event);
-    }
-
-    this._changeSelectionData = null;
-    if (current && current.value) {
-      const { keyCode, ctrlKey, metaKey } = event;
-
-      // Ignore ctrl and meta keydown
-      if (ctrlKey || metaKey) {
+  const handleInputChange = React.useCallback(
+    (ev: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>, inputValue: string) => {
+      if (internalState.changeSelectionData === null && textField.current) {
+        internalState.changeSelectionData = {
+          changeType: 'default',
+          selectionStart: textField.current.selectionStart !== null ? textField.current.selectionStart : -1,
+          selectionEnd: textField.current.selectionEnd !== null ? textField.current.selectionEnd : -1,
+        };
+      }
+      if (!internalState.changeSelectionData) {
         return;
       }
 
-      // On backspace or delete, store the selection and the keyCode
-      if (keyCode === KeyCodes.backspace || keyCode === KeyCodes.del) {
-        const selectionStart = (event.target as HTMLInputElement).selectionStart;
-        const selectionEnd = (event.target as HTMLInputElement).selectionEnd;
+      // The initial value of cursorPos does not matter
+      let cursorPos = 0;
+      const { changeType, selectionStart, selectionEnd } = internalState.changeSelectionData;
 
-        // Check if backspace or delete press is valid.
-        if (
-          !(keyCode === KeyCodes.backspace && selectionEnd && selectionEnd > 0) &&
-          !(keyCode === KeyCodes.del && selectionStart !== null && selectionStart < current.value.length)
-        ) {
+      if (changeType === 'textPasted') {
+        const charsSelected = selectionEnd - selectionStart;
+        const charCount = inputValue.length + charsSelected - displayValue.length;
+        const startPos = selectionStart;
+        const pastedString = inputValue.substr(startPos, charCount);
+
+        // Clear any selected characters
+        if (charsSelected) {
+          internalState.maskCharData = clearRange(internalState.maskCharData, selectionStart, charsSelected);
+        }
+        cursorPos = insertString(internalState.maskCharData, startPos, pastedString);
+      } else if (changeType === 'delete' || changeType === 'backspace') {
+        // isDel is true If the characters are removed LTR, otherwise RTL
+        const isDel = changeType === 'delete';
+        const charCount = selectionEnd - selectionStart;
+
+        if (charCount) {
+          // charCount is > 0 if range was deleted
+          internalState.maskCharData = clearRange(internalState.maskCharData, selectionStart, charCount);
+          cursorPos = getRightFormatIndex(internalState.maskCharData, selectionStart);
+        } else {
+          // If charCount === 0, there was no selection and a single character was deleted
+          if (isDel) {
+            internalState.maskCharData = clearNext(internalState.maskCharData, selectionStart);
+            cursorPos = getRightFormatIndex(internalState.maskCharData, selectionStart);
+          } else {
+            internalState.maskCharData = clearPrev(internalState.maskCharData, selectionStart);
+            cursorPos = getLeftFormatIndex(internalState.maskCharData, selectionStart);
+          }
+        }
+      } else if (inputValue.length > displayValue.length) {
+        // This case is if the user added characters
+        const charCount = inputValue.length - displayValue.length;
+        const startPos = selectionEnd - charCount;
+        const enteredString = inputValue.substr(startPos, charCount);
+
+        cursorPos = insertString(internalState.maskCharData, startPos, enteredString);
+      } else if (inputValue.length <= displayValue.length) {
+        /**
+         * This case is reached only if the user has selected a block of 1 or more
+         * characters and input a character replacing the characters they've selected.
+         */
+        const charCount = 1;
+        const selectCount = displayValue.length + charCount - inputValue.length;
+        const startPos = selectionEnd - charCount;
+        const enteredString = inputValue.substr(startPos, charCount);
+
+        // Clear the selected range
+        internalState.maskCharData = clearRange(internalState.maskCharData, startPos, selectCount);
+        // Insert the printed character
+        cursorPos = insertString(internalState.maskCharData, startPos, enteredString);
+      }
+
+      internalState.changeSelectionData = null;
+
+      const newValue = getMaskDisplay(mask, internalState.maskCharData, maskChar);
+
+      setDisplayValue(newValue);
+      setMaskCursorPosition(cursorPos);
+
+      // Perform onChange after input has been processed. Return value is expected to be the displayed text
+      onChange?.(ev, newValue);
+    },
+    [displayValue.length, internalState, mask, maskChar, onChange],
+  );
+
+  const handleKeyDown = React.useCallback(
+    (ev: React.KeyboardEvent<HTMLInputElement>) => {
+      onKeyDown?.(ev);
+
+      internalState.changeSelectionData = null;
+      if (textField.current && textField.current.value) {
+        const { keyCode, ctrlKey, metaKey } = ev;
+
+        // Ignore ctrl and meta keydown
+        if (ctrlKey || metaKey) {
           return;
         }
 
-        this._changeSelectionData = {
-          changeType: keyCode === KeyCodes.backspace ? 'backspace' : 'delete',
-          selectionStart: selectionStart !== null ? selectionStart : -1,
-          selectionEnd: selectionEnd !== null ? selectionEnd : -1,
-        };
+        // On backspace or delete, store the selection and the keyCode
+        if (keyCode === KeyCodes.backspace || keyCode === KeyCodes.del) {
+          const selectionStart = (ev.target as HTMLInputElement).selectionStart;
+          const selectionEnd = (ev.target as HTMLInputElement).selectionEnd;
+
+          // Check if backspace or delete press is valid.
+          if (
+            !(keyCode === KeyCodes.backspace && selectionEnd && selectionEnd > 0) &&
+            !(keyCode === KeyCodes.del && selectionStart !== null && selectionStart < textField.current.value.length)
+          ) {
+            return;
+          }
+
+          internalState.changeSelectionData = {
+            changeType: keyCode === KeyCodes.backspace ? 'backspace' : 'delete',
+            selectionStart: selectionStart !== null ? selectionStart : -1,
+            selectionEnd: selectionEnd !== null ? selectionEnd : -1,
+          };
+        }
       }
-    }
-  };
+    },
+    [internalState, onKeyDown],
+  );
 
-  private _onPaste = (event: React.ClipboardEvent<HTMLInputElement>) => {
-    if (this.props.onPaste) {
-      this.props.onPaste(event);
-    }
+  const handlePaste = React.useCallback(
+    (ev: React.ClipboardEvent<HTMLInputElement>) => {
+      onPaste?.(ev);
 
-    const selectionStart = (event.target as HTMLInputElement).selectionStart;
-    const selectionEnd = (event.target as HTMLInputElement).selectionEnd;
-    // Store the paste selection range
-    this._changeSelectionData = {
-      changeType: 'textPasted',
-      selectionStart: selectionStart !== null ? selectionStart : -1,
-      selectionEnd: selectionEnd !== null ? selectionEnd : -1,
-    };
-  };
-}
+      const selectionStart = (ev.target as HTMLInputElement).selectionStart;
+      const selectionEnd = (ev.target as HTMLInputElement).selectionEnd;
+      // Store the paste selection range
+      internalState.changeSelectionData = {
+        changeType: 'textPasted',
+        selectionStart: selectionStart !== null ? selectionStart : -1,
+        selectionEnd: selectionEnd !== null ? selectionEnd : -1,
+      };
+    },
+    [internalState, onPaste],
+  );
+
+  // Updates the display value if mask or value props change.
+  React.useEffect(() => {
+    internalState.maskCharData = parseMask(mask, maskFormat);
+    value !== undefined && setValue(value);
+    setDisplayValue(getMaskDisplay(mask, internalState.maskCharData, maskChar));
+    // eslint-disable-next-line react-hooks/exhaustive-deps -- Should only update when mask or value changes.
+  }, [mask, value]);
+
+  // Run before browser paint to avoid flickering from selection reset.
+  React.useLayoutEffect(() => {
+    // Move the cursor to position before paint.
+    if (maskCursorPosition !== undefined && textField.current) {
+      textField.current.setSelectionRange(maskCursorPosition, maskCursorPosition);
+    }
+  }, [maskCursorPosition]);
+
+  // Run after browser paint.
+  React.useEffect(() => {
+    // Move the cursor to the start of the mask format after values update.
+    if (internalState.isFocused && maskCursorPosition !== undefined && textField.current) {
+      textField.current.setSelectionRange(maskCursorPosition, maskCursorPosition);
+    }
+  });
+
+  useComponentRef(componentRef, internalState, textField);
+
+  return (
+    <TextField
+      {...props}
+      elementRef={ref}
+      onFocus={handleFocus}
+      onBlur={handleBlur}
+      onMouseDown={handleMouseDown}
+      onMouseUp={handleMouseUp}
+      onChange={handleInputChange}
+      onKeyDown={handleKeyDown}
+      onPaste={handlePaste}
+      value={displayValue || ''}
+      componentRef={textField}
+    />
+  );
+});
+MaskedTextField.displayName = COMPONENT_NAME;
