--- packages/react-next/src/components/TextField/index.ts
@@ -1,4 +1,0 @@
-export * from './TextField';
-export * from './TextField.base';
-export * from './TextField.types';
-export * from './MaskedTextField/MaskedTextField';
